HTCONDOR-2601 Consolidate networking functions
